use mulkit to do matrix dot, viktor can't do matrix dot, just vector

+++ Android/app/src/main/java/com/example/emojisemanticsearch/entity/EmojiJsonEntity.kt
@@ -3,7 +3,7 @@ package com.example.emojisemanticsearch.entity
 data class EmojiJsonEntity(
     val emoji: String,
     val message: String,
-    val embed: DoubleArray
+    val embed: FloatArray
 ) {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
